+++ corehq/apps/export/forms.py
@@ -941,7 +941,7 @@ class EmwfFilterFormExport(EmwfFilterExportMixin, GenericFilterFormExportDownloa
         return user_types
 
     def get_edit_url(self, export):
-        from corehq.apps.export.views import EditNewCustomFormExportView
+        from corehq.apps.export.old_views import EditNewCustomFormExportView
         return reverse(EditNewCustomFormExportView.urlname,
                        args=(self.domain_object.name, export._id))
 
@@ -1012,7 +1012,7 @@ class FilterCaseESExportDownloadForm(EmwfFilterExportMixin, BaseFilterExportDown
         )
 
     def get_edit_url(self, export):
-        from corehq.apps.export.views import EditNewCustomCaseExportView
+        from corehq.apps.export.old_views import EditNewCustomCaseExportView
         return reverse(EditNewCustomCaseExportView.urlname,
                        args=(self.domain_object.name, export.get_id))
